README and moving each backend to separate files
