+++ fang/Cargo.toml
@@ -52,7 +52,7 @@ typed-builder = "0.14"
 typetag = "0.2"
 uuid = { version = "1.1", features = ["v4"] }
 fang-derive-error = { version = "0.1.0" , optional = true}
-sqlx = {version = "0.6.3", features = ["any" , "macros" , "runtime-tokio-rustls", "postgres", "sqlite", "mysql"], optional = true}
+sqlx = {version = "0.6.3", features = ["any" , "macros" , "runtime-tokio-rustls"], optional = true}
 
 [dependencies.diesel]
 version = "2.1"
